+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 
+- Added a generic return type to the \ExtendsSoftware\ExaPHP\ServiceLocator\ServiceLocatorInterface::getService method.
+
 ### Changed
 
 ### Deprecated